refactor: extract date interval utilities
